+++ src/backendFile/aquapurebackend/src/main/java/com/backend/aquapurebackend/dto/ShopCartPostDTO.java
@@ -2,12 +2,12 @@ package com.backend.aquapurebackend.dto;
 
 
 public class ShopCartPostDTO {
-    int id;
-	int userid;
-	int product_id;
+    Long id;
+	Long userid;
+	Long product_id;
 	int quantity;
 	
-	public ShopCartPostDTO(int id, int userid, int product_id, int quantity) {
+	public ShopCartPostDTO(Long id, Long userid, Long product_id, int quantity) {
 		this.id = id;
 		this.userid = userid;
 		this.product_id = product_id;
@@ -15,27 +15,27 @@ public class ShopCartPostDTO {
 		
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
-	public int getUserid() {
+	public Long getUserid() {
 		return userid;
 	}
 
-	public void setUserid(int userid) {
+	public void setUserid(Long userid) {
 		this.userid = userid;
 	}
 
-	public int getProduct_id() {
+	public Long getProduct_id() {
 		return product_id;
 	}
 
-	public void setProduct_id(int product_id) {
+	public void setProduct_id(Long product_id) {
 		this.product_id = product_id;
 	}
 	public int getquantity() {
