+++ src/backendFile/aquapurebackend/src/main/java/com/backend/aquapurebackend/repository/ShopCartRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.repository.CrudRepository;
 import com.backend.aquapurebackend.model.ShopCart;
 
 public interface ShopCartRepository extends CrudRepository<ShopCart,Long> {
-    List<ShopCart> findByUserid(int userid);
-    List<ShopCart> deleteByUserid(int userid);
+    List<ShopCart> findByUserid(Long userid);
+    List<ShopCart> deleteByUserid(Long userid);
 }
