+++ src/backendFile/aquapurebackend/src/main/java/com/backend/aquapurebackend/repository/ShopPurchaseRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.repository.CrudRepository;
 import com.backend.aquapurebackend.model.ShopPurchase;
 
 public interface ShopPurchaseRepository extends CrudRepository<ShopPurchase,Long>{
-    List<ShopPurchase> findByUserid(int userid);
+    List<ShopPurchase> findByUserid(Long userid);
 }
